quickstart refactor
